# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brightloom API",
    "version": "1.0.0",
    "summary": "Tasks, projects and billing for small teams.",
    "description": "The HTTP API behind the Brightloom web app. Every endpoint apart from the login call and the billing webhook needs the session cookie that POST /api/v1/auth/login sets.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "contact": {
      "name": "Brightloom engineering",
      "email": "engineering@brightloom.test"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4100",
      "description": "Local development server"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "login",
        "summary": "Sign in with an email address and password",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The signed in user and organisation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "logout",
        "summary": "Sign out and clear the session cookie",
        "security": [],
        "responses": {
          "200": {
            "description": "The session was cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "getCurrentUser",
        "summary": "Read the signed in user and organisation",
        "responses": {
          "200": {
            "description": "The current user and organisation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "tags": [
          "Auth"
        ],
        "operationId": "updateCurrentUser",
        "summary": "Update the signed in user profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/org": {
      "get": {
        "tags": [
          "Organisation"
        ],
        "operationId": "getOrganisation",
        "summary": "Read the organisation and its members",
        "responses": {
          "200": {
            "description": "The organisation and its members",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgWithMembersResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "patch": {
        "tags": [
          "Organisation"
        ],
        "operationId": "updateOrganisation",
        "summary": "Rename the organisation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated organisation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrgResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/v1/projects": {
      "get": {
        "tags": [
          "Projects"
        ],
        "operationId": "listProjects",
        "summary": "List the projects in the organisation",
        "responses": {
          "200": {
            "description": "The organisation projects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "operationId": "createProject",
        "summary": "Create a project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Project id",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "patch": {
        "tags": [
          "Projects"
        ],
        "operationId": "updateProject",
        "summary": "Rename, recolour or archive a project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "operationId": "deleteProject",
        "summary": "Delete a project",
        "responses": {
          "200": {
            "description": "The project was deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/billing/subscription": {
      "get": {
        "tags": [
          "Billing"
        ],
        "operationId": "getSubscription",
        "summary": "Read the plan, seat count and next invoice",
        "responses": {
          "200": {
            "description": "The current subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/billing/invoices": {
      "get": {
        "tags": [
          "Billing"
        ],
        "operationId": "listInvoices",
        "summary": "List the invoices for the organisation",
        "responses": {
          "200": {
            "description": "The invoices, newest period first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/billing/webhook": {
      "post": {
        "tags": [
          "Billing"
        ],
        "operationId": "billingWebhook",
        "summary": "Receive a payment provider event",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The event was accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bl_session",
        "description": "The session cookie set by POST /api/v1/auth/login."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "The request was not valid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "No valid session cookie was sent",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The session may not perform this action",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "No record matches that id",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "examples": [
                  "not_found"
                ]
              },
              "message": {
                "type": "string",
                "examples": [
                  "That task does not exist"
                ]
              }
            }
          }
        }
      },
      "OkResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "email",
          "name",
          "role",
          "timezone"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "org_id": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ]
          },
          "timezone": {
            "type": "string",
            "examples": [
              "Europe/London"
            ]
          }
        }
      },
      "Member": {
        "type": "object",
        "required": [
          "id",
          "name",
          "email",
          "role"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "owner",
              "member"
            ]
          }
        }
      },
      "Org": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "plan"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "plan": {
            "type": "string",
            "enum": [
              "starter",
              "team",
              "enterprise"
            ]
          }
        }
      },
      "Project": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "name",
          "color",
          "archived",
          "created_at",
          "open_task_count"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "org_id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "enum": [
              "blue",
              "amber",
              "violet",
              "green",
              "rose",
              "slate"
            ]
          },
          "archived": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "open_task_count": {
            "type": "integer"
          }
        }
      },
...
      "Invoice": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "amount_cents",
          "status",
          "period",
          "issued_at"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "org_id": {
            "type": "integer"
          },
          "amount_cents": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "paid",
              "open",
              "void"
            ]
          },
          "period": {
            "type": "string",
            "examples": [
              "2026-08"
            ]
          },
          "issued_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "examples": [
              "priya@brightloom.test"
            ]
          },
          "password": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "SessionResponse": {
        "type": "object",
        "required": [
          "user",
          "org"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "org": {
            "$ref": "#/components/schemas/Org"
          }
        }
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "user"
        ],
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "ProfileUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string",
            "examples": [
              "Europe/London"
            ]
          }
        }
      },
      "OrgUpdateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "OrgResponse": {
        "type": "object",
        "required": [
          "org"
        ],
        "properties": {
          "org": {
            "$ref": "#/components/schemas/Org"
          }
        }
      },
      "OrgWithMembersResponse": {
        "type": "object",
        "required": [
          "org",
          "members"
        ],
        "properties": {
          "org": {
            "$ref": "#/components/schemas/Org"
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Member"
            }
          }
        }
      },
      "ProjectCreateRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "color": {
            "type": "string",
            "enum": [
              "blue",
              "amber",
              "violet",
              "green",
              "rose",
              "slate"
            ]
          }
        }
      },
      "ProjectUpdateRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "color": {
            "type": "string",
            "enum": [
              "blue",
              "amber",
              "violet",
              "green",
              "rose",
              "slate"
            ]
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
      "ProjectResponse": {
        "type": "object",
        "required": [
          "project"
        ],
        "properties": {
          "project": {
            "$ref": "#/components/schemas/Project"
          }
        }
      },
      "ProjectListResponse": {
        "type": "object",
        "required": [
          "projects"
        ],
        "properties": {
          "projects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Project"
            }
          }
        }
      },
...
      "SubscriptionResponse": {
        "type": "object",
        "required": [
          "plan",
          "seats",
          "seat_price_cents",
          "next_invoice"
        ],
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "starter",
              "team",
              "enterprise"
            ]
          },
          "seats": {
            "type": "integer"
          },
          "seat_price_cents": {
            "type": "integer"
          },
          "next_invoice": {
            "type": "object",
            "required": [
              "period",
              "amount_cents",
              "due_date"
            ],
            "properties": {
              "period": {
                "type": "string",
                "examples": [
                  "2026-10"
                ]
              },
              "amount_cents": {
                "type": "integer"
              },
              "due_date": {
                "type": "string",
                "examples": [
                  "2026-10-01"
                ]
              }
            }
          }
        }
      },
      "InvoiceListResponse": {
        "type": "object",
        "required": [
          "invoices"
        ],
        "properties": {
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invoice"
            }
          }
        }
      },
      "WebhookEvent": {
        "type": "object",
        "required": [
          "type"
        ],
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "subscription.updated",
              "invoice.paid"
            ]
          },
          "org_slug": {
            "type": "string",
            "examples": [
              "brightloom"
            ]
          },
          "plan": {
            "type": "string",
            "enum": [
              "starter",
              "team",
              "enterprise"
            ]
          },
          "invoice_id": {
            "type": "integer"
          }
        }
      },
      "WebhookResponse": {
        "type": "object",
        "required": [
          "received"
        ],
        "properties": {
          "received": {
            "type": "boolean"
          }
        }
      }
    }
  }
}